+++ apps/rig_inbound_gateway/test/rig_web/presence/channel_test.exs
@@ -3,6 +3,15 @@ defmodule RigInboundGatewayWeb.Presence.ChannelTest do
   use ExUnit.Case, async: true
   import ExUnit.CaptureLog
   use RigInboundGatewayWeb.ChannelCase
+  use RigInboundGatewayWeb.ConnCase
+
+  setup do
+    conn =
+      build_conn()
+      |> put_req_header("content-type", "application/json")
+
+    {:ok, conn: conn}
+  end
 
   test "a user connecting to her own topic works" do
     assert {:ok, _response, sock} = subscribe_and_join_user(
@@ -51,5 +60,40 @@ defmodule RigInboundGatewayWeb.Presence.ChannelTest do
       )
     end
     assert capture_log(fun) =~ "unauthorized"
+  end
+
+  @tag :smoke
+  test "subscribed user should receive message", %{conn: conn} do
+    {:ok, _msg} = RigOutboundGateway.Kafka.GroupSubscriber.wait_for_consumer_ready("rig", 0)
+    assert {:ok, _response, sock} = subscribe_and_join_user(
+      "testuser",
+      [@customer_role],
+      "user:testuser"
+    )
+
+    body = ~s({"user":"testuser","foo":"bar"})
+    produce_kafka_message(conn, body)
+
+    expected_event = "message"
+    expected_payload = %{"user" => "testuser", "foo" => "bar"}
+    assert_broadcast ^expected_event, ^expected_payload, 3000
+
+    leave sock
+  end
+
+  @tag :smoke
+  test "not subscribed user shouldn'\t receive message", %{conn: conn} do
+    {:ok, _msg} = RigOutboundGateway.Kafka.GroupSubscriber.wait_for_consumer_ready("rig", 0)
+    body = ~s({"user":"testuser","foo":"bar"})
+    produce_kafka_message(conn, body)
+
+    expected_event = "message"
+    expected_payload = %{"user" => "testuser", "foo" => "bar"}
+    refute_broadcast ^expected_event, ^expected_payload, 2000
+  end
+
+  defp produce_kafka_message(conn, body) do
+    conn = post conn, "/kafka/produce", body
+    assert conn.status == 200
   end
 end