+++ apps/rig_inbound_gateway/test/test_helper.exs
@@ -1,2 +1,4 @@
 ExUnit.start()
+# Exclude all smoke tests from running by default
+ExUnit.configure exclude: [smoke: true]
 Application.ensure_all_started(:bypass)